move WorkflowRunDetail into parent page for cleaner code
