+++ app/workflow-runs/[traceId]/page.tsx
@@ -4,10 +4,54 @@ import { notFound } from "next/navigation"
 
 import BaseGitHubItemCard from "@/components/github/BaseGitHubItemCard"
 import { Button } from "@/components/ui/button"
-import WorkflowRunDetail from "@/components/workflow-runs/WorkflowRunDetail"
+import {
+  DefaultEvent,
+  ErrorEvent,
+  LLMResponseEvent,
+  StatusUpdate,
+  SystemPromptEvent,
+  ToolCallEvent,
+  ToolResponseEvent,
+  UserMessageEvent,
+} from "@/components/workflow-runs/events"
 import { getIssue } from "@/lib/github/issues"
+import { n4j } from "@/lib/neo4j/service"
 import { WorkflowPersistenceService } from "@/lib/services/WorkflowPersistenceService"
 import { WorkflowWithEvents } from "@/lib/types/workflow"
+import { WorkflowEvent } from "@/lib/types/workflow"
+
+function EventContent({
+  event,
+  issue,
+}: {
+  event: WorkflowEvent
+  issue?: { number: number; repoFullName: string }
+}) {
+  switch (event.type) {
+    case "status":
+      return <StatusUpdate event={event} timestamp={event.timestamp} />
+    case "system_prompt":
+      return <SystemPromptEvent event={event} timestamp={event.timestamp} />
+    case "llm_response":
+      return (
+        <LLMResponseEvent
+          event={event}
+          timestamp={event.timestamp}
+          issue={issue}
+        />
+      )
+    case "user_message":
+      return <UserMessageEvent event={event} timestamp={event.timestamp} />
+    case "tool_call":
+      return <ToolCallEvent event={event} />
+    case "tool_response":
+      return <ToolResponseEvent event={event} />
+    case "error":
+      return <ErrorEvent event={event} />
+    default:
+      return <DefaultEvent event={event} />
+  }
+}
 
 export default async function WorkflowRunDetailPage({
   params,
@@ -16,6 +60,9 @@ export default async function WorkflowRunDetailPage({
 }) {
   const { traceId } = params
 
+  // const workflow = await n4j.getWorkflow(workflowId)
+  // const events = await n4j.listEventsForWorkflow(workflowId)
+  // const issue = await n4j.getIssueFromWorkflow(workflowId)
   const workflow: WorkflowWithEvents | null =
     await new WorkflowPersistenceService()
       .getWorkflowEvents(traceId)
@@ -78,10 +125,18 @@ export default async function WorkflowRunDetailPage({
         )}
 
         {/* Timeline Section */}
-        <div className="space-y-3">
-          <h2 className="text-lg font-semibold">Timeline</h2>
-          <WorkflowRunDetail events={workflow.events} issue={workflow.issue} />
-        </div>
+        {workflow.events && (
+          <div className="space-y-3">
+            <h2 className="text-lg font-semibold">Timeline</h2>
+            <div className="bg-card border rounded-lg overflow-hidden">
+              {workflow.events.map((event) => (
+                <div key={event.id} className="p-3 sm:p-4">
+                  <EventContent event={event} issue={issue} />
+                </div>
+              ))}
+            </div>
+          </div>
+        )}
       </div>
     </main>
   )
